fix: no-underscore-dangle __REDUX_DEVTOOLS_EXTENSION__
