--- src/utils/initialState.js
@@ -1,9 +1,0 @@
-const initialState = {
-  users: [],
-  pageSize: 5,
-  totalUsersCount: 0,
-  currentPage: 1,
-  isFetching: false,
-};
-
-export { initialState };
